// src/types/General/projects.ts
export interface Project {
  id: number;
  active?: boolean;
  name?: string;
  company_id?: number;
  company_name: number;
  project_manager_id?: number;
  project_name?: string;
  project_code?: string;
  estimated_start_date?: string;
  estimated_completion_date?: string;
  deleted?: boolean;
  profit_share?: number;
  weblink_to_layout_plan?: string;
  property_id?: number;
  sub_property_id?: number;
  land_area?: number;
  land_cost?: number;
  saleable_area?: number;
  proposed_usage?: string;
  street?: string;
  city?: string;
  state?: string;
  country?: string;
  pin?: string;
  project_address_line: string;
  same_as_property_address?: boolean;
  project_address?: string;
  number_of_blocks?: number;
  total_no_of_saleable_units?: number;
  common_garden_area?: number;
  promotion_budget_per_sq_ft?: number;
  budget_amount?: number;
  promotion_budget_amount?: number;
  advance_interest_rate?: number;
  delayed_interest_penalty_rate?: number;
  swift_account_details?: string;
  project_amenities?: string;
  additional_design_highlights?: string;
  location_summary?: string;
  architect?: string;
  nature_of_work?: string;
  sub_project?: string;
  estimated_carpet_area?: number;
  property_manager_id?: number;
  same_as_client_master?: boolean;
  location?: string;
  type_of_project?: string;
  certifying_company_id?: number;
  certifying_company_gstin_id?: number;
  lease_certifying_company_id?: number;
  lease_certifying_company_gstin_id?: number;
  pms_supplier_id?: number;
  client_address?: string;
  construction_area?: number;
  project_address_attributes?: {
    id?: number;
    address?: string;
    address_type?: string;
    address_line_two?: string;
    pms_country_id?: number;
    pms_state_id?: number;
    pms_city_id?: number;
    pms_location_id?: number;
    pin_code?: string;
    telephone_number?: string;
    fax_number?: string;
    _destroy?: boolean;
  };

  property_address_attributes?: {
    id?: number;
    address?: string;
    address_type?: string;
    address_line_two?: string;
    pms_country_id?: number;
    pms_state_id?: number;
    pms_city_id?: number;
    pms_location_id?: number;
    pin_code?: string;
    telephone_number?: string;
    fax_number?: string;
    _destroy?: boolean;
  };

  project_architect_details_attributes?: Array<{
    id?: number;
    architect?: string;
    nature_of_work?: string;
    _destroy?: boolean;
  }>;
}