// src/types/General/sites.ts
export interface Site {
  id: number;
  seleable_area?: number;
  development_type_id?: number;
  common_tarrace_area?: number;
  alloted_land_area?: number;
  rera_area?: number;
  name?: string;
  project_id?: number;
  company_id?: number;
  region_id?: number;
  user_id?: number;
  headquarter_id?: number;
  latitude?: number;
  longitude?: number;
  visitor_image_required?: boolean;
  visitor_identification_required?: boolean;
  attendance_image_required?: boolean;
  geofence_range?: number;
  zone_id?: number;
  complaint_urgency_enabled?: boolean;
  custom_fields?: unknown;
  branch_code?: string;
  ifsc_code?: string;
  tier?: string;
  classification?: string;
  address?: string;
  city?: string;
  district?: string;
  state?: string;
  pmc_vendor_id?: number;
  location: object;
  location_url?: string;
  temp_mandatory?: boolean;
  health_questions_enabled?: boolean;
  temp_scan_enabled?: boolean;
  fr_enabled?: boolean;
  arogya_setu_scanner_enabled?: boolean;
  inv_online_payment_allowed?: boolean;
  auto_create_invoice?: boolean;
  auto_create_receipt?: boolean;
  invoice_format?: string;
  receipt_format?: string;
  invoice_prefix?: string;
  invoice_next_number?: number;
  receipt_prefix?: string;
  receipt_next_number?: number;
  bills_and_payments_enabled?: boolean;
  ccavenue_sub_account_id?: string;
  skip_host_approval?: boolean;
  registration_workflow?: string;
  restaurants_enabled?: boolean;
  facility_bookings_enabled?: boolean;
  no_of_devices?: number;
  device_rental_type?: string;
  device_rental_rate?: number;
  space_enabled?: boolean;
  max_device_limit?: number;
  survey_enabled?: boolean;
  fitout_area_rate?: number;
  fitout_enabled?: boolean;
  mailroom_enabled?: boolean;
  create_breakdown_ticket?: boolean;
  parking_enabled?: boolean;
  default_visitor_pass?: string;
  ecommerce_service_enabled?: boolean;
  steps_enabled?: boolean;
  transportation_enabled?: boolean;
  business_card_enabled?: boolean;
  visitor_enabled?: boolean;
  sap_plant_code?: string;
  sites_start_dates?: string;
  total_area_sq_ft?: number;
  operational_audit_enabled?: boolean;
  patrolling_notification_time?: string;
  patrolling_approver_id?: number;
  project?: string;
  type_of_project?: string;
  sub_project_name?: string;
  company_name?: string;
  sub_project_code?: string;
  company_gstin_id?: number;
  lease_certifying_company_id?: number;
  accounting_package?: string;
  lease_certifying_company_gstin_id?: number;
  development_type?: string;
  estimated_start_date?: string;
  estimated_end_date?: string;
  possesion_date?: string;
  number_of_floor?: number;
  number_of_basement?: number;
  number_of_seleable_unit?: number;
  sales_uom?: string;
  registration_number?: string;
  registration_date?: string;
  oc_obtained?: boolean;
  oc_date?: string;
  oc_number?: string;
  sales_incentive?: string;
  rate_per_sq_ft_incentive_amount?: number;
  show_hold_unit_in_sales?: boolean;
  numer_of_car_park_mendatory?: number;
  plan_level_discount?: boolean;
  transaction_level_discount?: boolean;
  payee_ac_name?: string;
  include_in_base_cost_unit?: string[]; // array
  consider_for_brokerage?: string[]; // array
  include_in_agreement_value?: string[]; // array
  tds_applicable_on?: string[]; // array
  auto_unblocking?: string[]; // array
  blocking_period_in_hour?: number;
  rental_blocking_period_in_hour?: number;
  allow_unit_blocking_without_money?: boolean;
  plc_discount_app?: boolean;
  auto_tax_cal?: boolean;
  escrow_project?: boolean;
  is_common_service?: boolean;
  budget_type?: string;
  eng_uom?: string;
  eng_area?: number;
  const_area?: number;
  budget_cal_type?: string;
  cons_budget_sq_ft?: number;
  cons_budget_sq_mt?: number;
  budget_revenue_sq_ft?: number;
  add_int?: boolean;
  delayed_interest_penalty_due?: boolean;
  delayed_interest_penalty_rate?: number;
  delayed_interest_penalty_grace_pe?: number;
  ad_int_type?: string;
  ad_int_rate?: number;
  sub_project?: string;
  deleted?: boolean;
  budget_amount?: number;
  effective_date?: string;
  podium_level?: number;
  still_level?: number;
  status?: string;
  remark?: string;
  active?: boolean;

  site_intrest_rates_attributes?: Array<{
    id?: number;
    from_days?: number;
    to_days?: number;
    rate?: number;
    _destroy?: boolean;
  }>;

  interest_sections_attributes?: Array<{
    id?: number;
    effectiv_date?: string;
    delayed_interest_penalty_due?: boolean;
    delayed_interest_penalty_grace_pe?: number;
    ad_int_type?: string;
    ad_int_rate?: number;
    _destroy?: boolean;
    site_intrest_rates_attributes?: Array<{
      id?: number;
      from_days?: number;
      to_days?: number;
      rate?: number;
      _destroy?: boolean;
    }>;
  }>;
}